requirements.txt
efficientnet_pytorch==0.7.1
h5py==2.10.0
keras==2.11.0
numpy==1.23.4
opencv_python==4.7.0.72
opencv_python_headless==4.6.0.66
pandas==1.5.1
Pillow==10.0.1
pretrainedmodels==0.7.4
pydicom==2.3.1
scikit_image==0.16.2
scikit_learn==1.3.1
scipy==1.8.1
SimpleITK==2.3.0
tensorflow==2.10.0
tensorflow_gpu==2.11.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
